server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
chromadb-client==0.5.18
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.5
frozenlist==1.5.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
marshmallow==3.23.1
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
orjson==3.10.11
overrides==7.7.0
packaging==24.2
posthog==3.7.0
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.17.1
python-multipart
zipp==3.21.0
